+++ packages/iov-bns-governance/types/proposals.d.ts
@@ -13,6 +13,7 @@ export declare enum ProposalType {
   ReleaseGuaranteeFunds = 6,
   DistributeFunds = 7,
   AmendProtocol = 8,
+  ExecuteMigration = 9,
 }
 interface CommonProposalOptions {
   readonly type: ProposalType;
@@ -66,6 +67,10 @@ export interface AmendProtocolOptions extends CommonProposalOptions {
   readonly type: ProposalType.AmendProtocol;
   readonly text: string;
 }
+export interface ExecuteMigrationOptions extends CommonProposalOptions {
+  readonly type: ProposalType.ExecuteMigration;
+  readonly id: string;
+}
 export declare type ProposalOptions =
   | AddCommitteeMemberOptions
   | RemoveCommitteeMemberOptions
@@ -75,5 +80,6 @@ export declare type ProposalOptions =
   | RemoveValidatorOptions
   | ReleaseGuaranteeFundsOptions
   | DistributeFundsOptions
-  | AmendProtocolOptions;
+  | AmendProtocolOptions
+  | ExecuteMigrationOptions;
 export {};