+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - @iov/bns: Add optional payer parameter to `BnsConnection.withDefaultFee`.
 - @iov/bns: `BnsConnection.searchTx` now resolves to an array of
   `ConfirmedAndSignedTransaction<UnsignedTransaction> | FailedTransaction`.
+- @iov/bns-governance: Add proposal type `ExecuteMigration`.
 - @iov/lisk: Export `pubkeyToAddress`.
 
 Breaking changes